Generate initial CEO scaffold and first test
This commit contains the scaffold generated structure for the CEO model.
Note that we've already goofed by naming the model "Ceo" and not "CEO",
but convention tends to dictate that only the first letter is
capitalized.  So, The model is "Ceo" and we want to call it "CEO".  We
will have to grow.

The first test using MiniTest spec has been created as well.