Get first Cucumber Scenario to work
This started with adding Cucumber and wanting to us it to drive the
application development.  This project is so open ended that having the
Cucumber framework seems helpful.  Also, Kristin thought it was a good
idea.  :-)

Cucumber requires some web browser emulation to report on the presences
of elements on the rendered page.  Therefore Capybara was added.

Cucumber is invoked primarily by Guard, but I left the default profiles
in the cucumber.yml file so that it could also be invoked from the
command line.

I also added guard-haml so that changes to the haml files triggers
guard.