Add test in for view content, but Step broke
After the last checkin, something quickly broke.
It looks like the view file was empty for the Cucumber
Step, even though the fixture was loading the data.

Configured Cucumber to use the existing fixtures and modified the
ceo.yml fixture to have unique names.  Got MatchControllerTest to verify
content (note that output is very rudimentary and brittle), but Cucumber
is still reporting that the controller variable is nil because the HAML
is trying 'nil.each'.  Maybe restore Ceo.create statements in stepdef?