MatchesController new, not show & match owns win
The MatchesController is more RESTful.  A Match is between two CEO's and
there is a winner when someone votes.  Therefore, the page to show the
two CEOs in a match is not the "show" page -- that shows the __results__
of a match -- it is the "new" page.  Therefore, updated the default root
to reflect that.

When someone submits their vote, then they are creating a Match record
-- the two CEOs and the winner.  I chose to simplify and have only one
winner option -- "Like".  This is a change from the conception of being
able to either "Like"  or "Hate" a CEO.  We may return the downvote
option in the future.

I had also planned on having a Ceo#win method, which would award a "win"
to a CEO.  However, since only matches can be won, that concept belongs
on the Match object.  There is a method on the Ceo model to get the
number of 'wins' and, when needed, we can create a scope on the Match
object for the leaders.

We will need to solve the question of how to select CEOs to go up
against one another at some point soon.

This is an interim checking, we still need to make the feature green by
getting the right flash message on the page.